Move confusingly named and located solvers.py
