+++ idaes/util/tests/test_download_bin.py
@@ -1,5 +1,5 @@
 import idaes
-import idaes.solvers
+import idaes.util.download_bin
 import pytest
 import os
 import idaes.logger as idaeslog
@@ -18,7 +18,7 @@ def test_dl_bin():
     _del_data_file(os.path.join(idaes.testing_directory, "version_lib.txt"))
     _del_data_file(os.path.join(idaes.testing_directory, "version_solvers.txt"))
     ll = _log.getEffectiveLevel() # verbose will set level to DEBUG
-    idaes.solvers.download_binaries(
+    idaes.util.download_bin.download_binaries(
         release=idaes.config.default_binary_release,
         verbose=True,
         to_path="testing")
@@ -31,7 +31,7 @@ def test_dl_bin_unknown():
     _del_data_file(os.path.join(idaes.testing_directory, "version_lib.txt"))
     _del_data_file(os.path.join(idaes.testing_directory, "version_solvers.txt"))
     with pytest.raises(Exception):
-        idaes.solvers.download_binaries(
+        idaes.util.download_bin.download_binaries(
             platform="unknown platform",
             release=idaes.config.default_binary_release,
             to_path="testing")
